refactor(native): isolate archiver command construction
